Image previews
Additionally to enable the preview feature, one must add it to
config.json:
```json
    "image_preview": {}
```
See documentation for additional settings - they shouldn't be necessary
in most cases.

A feature flag for sixel support is added.

Windows is basically unsupported, but it is possible to force the
feature to use half-blocks, if the `protocol` option is set with `type`
and `font_size`.