fix: image preview offset
There was a hardcoded offset of `1` for the borders. With @ulyssa's fix
for vertical splits, that can be removed now, as it diplaces all
previews one character too much to the right.

There some other offset bugs related to the layout-by-width in
`get_render_format` not being replicated by `line_preview`, which should
probably by unified some way.