fix not showing displaynames of already synced rooms
Fixes #149

The display_names were only populated from the
OriginalSyncRoomMemberEvent. When opening an already synced room,
display_names is not populated.

The ChatStore's `need_load` is changed to a hashmap with OwnedRoomIds as
keys and a bitmask of "needs" as value. The "need" thus can be `MESSAGES`
or `MEMBERS`. The task that loads messages now also loads members, if
that flag is set for the room, which is done when opening a window.

Thanks to <benjamin@computer.surgery> for finding the source of the bug
in that issue.