fix truncation/padding for non-ascii sender names
The previous implementation was trying to pad/truncate sender names to
28 columns by taking the number of bytes in the utf-8 encoding. This
overestimates the display width and can crash iamb when trying to
truncate in the middle of a codepoint.

For the unicode-aware implementation, I used the existing take_width
function. The original implementation of that function had a couple
bugs, which are fixed here.